backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
datasets
scikit-learn
matplotlib
seaborn 